mds: Support up to 4095 nodes [#1613]
Support up to 4095 nodes in the flat addressing scheme for TIPC, by encoding the
slot ID in the lower eight bits and the ones' complement of the subslot ID in
bits 8 to 11 in the node identifier of the TIPC address. The reason for taking
the ones' complement of the subslot ID is backwards compatibility with existing
installations, so that this enhancement can be upgraded in-service.