plm: disallow activation of blade if it is operationally disabled [#1751] The issue is that a Blade that is disabled is allowed to activate.
PICMG 3.0 states that shelf managers should deactivate a FRU if it has a
non-recoverable sensor assertion. After deactivating the FRU, some shelf
managers will attempt to reactivate the FRU.

PLM should disallow activation of a FRU that is operationally disabled. The
REPAIRED admin operation should be used if the admin wants to activate the
FRU.
if isolation fails ReadinessImpact should return AIS_OK (as it is
currently doing), and isolate-pending flag is set. There is no ERR_DEPLOYMENT return
code for ReadinessImpact. As long as we prevent an HE from activating while in
disabled state.

Testing Commands:
-----------------
1) set a non-recoverable sensor, so that the shelf manager will power down the FRU (you
   can probably simulate this by initiating extraction pending)
2) before the deactivation completes, call ReadinessImpact with RI_FAILURE
3) if simulating, once the blade has deactivated, attempt to activate it again


Testing, Expected Results:
--------------------------
1) blade should remain in inactive state with no readiness flags set
2) attempts to activate the blade, should result in deactivations