amfd: allow lock and unlock operation on NoRed MW SI. [#1834]
In the reported issue, amfd crashes during deletion of MW NoRed SI while
standby SC is coming up.

Here requirement is to bring down a payload node and delete its related configuration
like node, MW SI, MW SU etc. As per AMF PR doc section 7.1.4 , SI must be locked before
deleting it. Also AMF allows deletion of SI in unlocked state if it is unassigned, but
it is not the recommended way. Since lock operation on NoRed MW SI is not allowed,
the only way to delete is when it is unassigned. This imposes another criterion to bring
down the node or lock NoRed SU so that its MW SI gets unassigned and its deletion can proceed.
Even in this case also, AMF can pick the same SI and assign it to some other unassigned node
or any node joining the cluster that time. Thus there is no gaurantee that SI will remain
unassigned.

Patch allows lock and unlock admin op on MW NoRed SI accept the one assigned on
active SC.