log: fix incorrect usage of saImmRtObjectDelete on cfg app stream [#1330]
saImmRtObjectDelete() was always called when closing app streams,
even the stream was configurable. If it was the case, that IMM API
would return SA_AIS_ERR_NOT_EXIST and had WA in syslog.

This patch introduce one flag in `log_stream_t`, showing
the closed stream is runtime or configurable. And only
call to the IMM API if the app stream is runtime.
And this flag value is cached locally at SCs node.