AMFND: Resend oper_state message after headless for escalation continuation [#1902]
The oper_state message will be buffered in AMFND during headless period.
The order of this oper_state message as well as su_si_assign message
have being added to queue is important. When AMFD come backs, AMFND will
reassigns new msg_id for these buffered messages. The relative order
between these buffered messages needs to be maintained as the same as they
were firstly added, so that the sequence will be executed as the same as
in non-headless.