amfnd: avoid sending multiple node-switchover recovery request for same fault[#1935]
In the reported problem, AMFND sends multiple mode-switchover recovery request
for same fault when nodeautorepair is disabled. Sometimes, amfnd is successful
in sending the request but it may crash also.

In the reported problem, AMFND reports node-switchover recovery to AMFD. AMFD
removes assignments gracefully and does not reboot the node as nodeautorepair
is disabled. After this user deletes the faulty SU after perofrming lock and lock-in
admin operations. Same SU is added again and instantiated. AMFND while instantiating
the SU, sends recovery request to AMFD again for the same fault.

Patch resets failed_su in avnd_cb so that recovery request is not sent
for same fault again.