clm: Send MDS messages asynchronously to avoid blocking the main thread [#2220]
Send the join request message asynchronously to the CLM server. This avoids
blocking the CLM main thread for a potentially very long time in case no reply
is received (e.g. because of message loss or because the CLM server died before
being able to send the reply). The result is a faster recovery from such error
scenarios.