AMFD: Prioritize to failover absent assignment [#2161]
In non-headless scenario, if there is a SU becoming OUT-OF-SERVICE
due to node reboot, the situation should be handled in node_fail_su_oper,
where the assignment of OUT-OF-SERVICE SU will be removed. In other words,
the susi_success_su_oper is not supposed to handle an OUT-OF-SERVICE SU
but still having assignment. When susi_success_su_oper() is called, SU
having assignment must be IN-SERVICE.

In scenario of headless recovery, which currently outweights pending
assignment than absent assignment, that would cause susi_success_su_oper()
encounter an OUT-OF-SERVICE SU but still has assignment (even absent
assignment though). The result eventually is that AMFD will generate
unexpected assignment messages.

A solution is that AMFD can revert the order of headless recovery, which
prioritizes to failover absent assignment first. This change should also
be working, because it has already been supported in non-headless, where
a node restarts while assignment on another node is still in progress.